login log out with jwt and expiray done

+++ controllers/user/user.js
@@ -24,13 +24,25 @@ const createUser = async (req, res, next) => {
     return res.status(500).json({ "error": error })
   }
 };
+
 const getAllUsers = async (req, res) => {
-  await User.find({}).then(ele => {
-    res.send(ele)
-  }).catch(err => {
-    res.send(err)
-  })
+  try {
+    const isAdmin = await User.exists({ email: req.user.data, authToken: req.token, isAdmin: true })
+  if (isAdmin) {
+    await User.find({}).then(ele => {
+      res.send(ele)
+    }).catch(err => {
+      res.send(err)
+    })
+  }
+  else {
+    res.status(401).send("Forbidden")
+  }
+  } catch (error) {
+    res.status(500).json({ Msg: error })
+  }
 };
+
 const getUserById = async (req, res) => {
   try {
     const { userId } = req.params;
@@ -43,11 +55,17 @@ const getUserById = async (req, res) => {
 };
 const updateUser = async (req, res) => {
   try {
-    const updateUser = req.body
-    const isUserExits = await User.exists({ id: userId })
-    if (!isUserExits) res.status(400).json({ Msg: "User does not exists" })
-    var isUserUpdated = await User.findOneAndUpdate({ id: userId.replace(":", "") }, updateUser)
-    res.send()
+    const isAdmin = await User.exists({ email: req.user.data, authToken: req.token, isAdmin: true })
+    if (isAdmin) {
+      const updateUser = req.body
+      const isUserExits = await User.exists({ id: userId })
+      if (!isUserExits) res.status(400).json({ Msg: "User does not exists" })
+      var isUserUpdated = await User.findOneAndUpdate({ id: userId.replace(":", "") }, updateUser)
+      res.send()
+    }
+    else {
+      res.status(401).send("Forbidden")
+    }
   } catch (error) {
     res.status(500).json({ Msg: error })
   }
@@ -59,11 +77,17 @@ const deleteUser = async (req, res) => {
     if (!errors.isEmpty()) {
       return res.status(422).json({ errors: errors.array() })
     }
-    const { userId } = req.body;
-    const isUserExits = await User.exists({ id: userId })
-    if (!isUserExits) res.status(400).json({ Msg: "User does not exists" })
-    const isUserRemoved = await User.findOneAndRemove({ id: userId })
-    res.status(200).send(isUserRemoved)
+    const isAdmin = await User.exists({ email: req.user.data, authToken: req.token, isAdmin: true })
+    if (isAdmin) {
+      const { userId } = req.body;
+      const isUserExits = await User.exists({ id: userId })
+      if (!isUserExits) res.status(400).json({ Msg: "User does not exists" })
+      const isUserRemoved = await User.findOneAndRemove({ id: userId })
+      res.status(200).send(isUserRemoved)
+    }
+    else {
+      res.status(401).send("Forbidden")
+    }
   } catch (error) {
     res.status(500).json({ Msg: error })
   }
